+++ yash-syntax/src/source/pretty.rs
@@ -24,7 +24,7 @@
 //! diagnostic message string.
 //!
 //! When the `yash_syntax` crate is built with the `annotate-snippets` feature
-//! enabled, it supports conversion from `Message` to `Snippet`. If you would
+//! enabled, it supports conversion from `Message` to `Group`. If you would
 //! like to use another formatter instead, you can provide your own conversion
 //! for yourself.
 //!
@@ -45,8 +45,8 @@
 //! // The lines below require the `annotate-snippets` feature.
 //! # #[cfg(feature = "annotate-snippets")]
 //! # {
-//! let message = annotate_snippets::Message::from(&message);
-//! eprint!("{}", annotate_snippets::Renderer::plain().render(message));
+//! let group = annotate_snippets::Group::from(&message);
+//! eprint!("{}", annotate_snippets::Renderer::plain().render(&[group]));
 //! # }
 //! ```
 //!
@@ -272,29 +272,39 @@ mod annotate_snippets_support {
     ///
     /// This implementation is only available when the `yash_syntax` crate is
     /// built with the `annotate-snippets` feature enabled.
-    impl From<AnnotationType> for annotate_snippets::Level {
+    impl<'a> From<AnnotationType> for annotate_snippets::Level<'a> {
         fn from(r#type: AnnotationType) -> Self {
             use AnnotationType::*;
             match r#type {
-                Error => Self::Error,
-                Warning => Self::Warning,
-                Info => Self::Info,
-                Note => Self::Note,
-                Help => Self::Help,
+                Error => Self::ERROR,
+                Warning => Self::WARNING,
+                Info => Self::INFO,
+                Note => Self::NOTE,
+                Help => Self::HELP,
+            }
+        }
+    }
+
+    impl From<AnnotationType> for annotate_snippets::AnnotationKind {
+        fn from(r#type: AnnotationType) -> Self {
+            use AnnotationType::*;
+            match r#type {
+                Error | Warning => Self::Primary,
+                Info | Note | Help => Self::Context,
             }
         }
     }
 
     /// Converts `yash_syntax::source::pretty::Message` into
-    /// `annotate_snippets::Message`.
+    /// `annotate_snippets::Group`.
     ///
     /// This implementation is only available when the `yash_syntax` crate is
     /// built with the `annotate-snippets` feature enabled.
-    impl<'a> From<&'a Message<'a>> for annotate_snippets::Message<'a> {
+    impl<'a> From<&'a Message<'a>> for annotate_snippets::Group<'a> {
         fn from(message: &'a Message<'a>) -> Self {
             let mut snippets: Vec<(
                 &super::super::Code,
-                annotate_snippets::Snippet,
+                annotate_snippets::Snippet<'a, annotate_snippets::Annotation<'a>>,
                 Vec<annotate_snippets::Annotation>,
             )> = Vec::new();
             // We basically convert each annotation into a snippet, but want to merge annotations
@@ -303,8 +313,9 @@ mod annotate_snippets_support {
             // snippet.
             for annotation in &message.annotations {
                 let range = annotation.location.byte_range();
-                let level = annotate_snippets::Level::from(annotation.r#type);
-                let as_annotation = level.span(range).label(&annotation.label);
+                let as_annotation = annotate_snippets::AnnotationKind::from(annotation.r#type)
+                    .span(range)
+                    .label(&annotation.label);
                 let code = &*annotation.location.code;
                 if let Some((_, _, annotations)) =
                     snippets.iter_mut().find(|&&mut (c, _, _)| c == code)
@@ -316,24 +327,22 @@ mod annotate_snippets_support {
                         .get()
                         .try_into()
                         .unwrap_or(usize::MAX);
-                    let snippet = annotate_snippets::Snippet::source(&annotation.code)
+                    let snippet = annotate_snippets::Snippet::source(&**annotation.code)
                         .line_start(line_start)
-                        .origin(code.source.label())
-                        .fold(true);
+                        .path(code.source.label());
                     snippets.push((code, snippet, vec![as_annotation]));
                 }
             }
 
             annotate_snippets::Level::from(message.r#type)
-                .title(&message.title)
-                .snippets(
+                .primary_title(&*message.title)
+                .elements(
                     snippets
                         .into_iter()
                         .map(|(_, snippet, annotations)| snippet.annotations(annotations)),
                 )
-                .footers(message.footers.iter().map(|footer| {
-                    let level = annotate_snippets::Level::from(footer.r#type);
-                    level.title(&footer.label)
+                .elements(message.footers.iter().map(|footer| {
+                    annotate_snippets::Level::from(footer.r#type).message(&*footer.label)
                 }))
         }
     }